requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.15.1
djangorestframework-stubs==3.14.5
idna==3.7
mypy==1.7.1
mypy-extensions==1.0.0
requests==2.31.0
sqlparse==0.5.0
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240406
typing_extensions==4.11.0
urllib3==2.2.1
